勘误信息bug，已修改

+++ src/main/java/po/ErrorInfo.java
@@ -10,7 +10,7 @@ public class ErrorInfo {
 	private int state;  //审核状态
 	private int verifyAdministratorId;//审核勘误的管理员
 	private int userId;//提交勘误的用户
-	private int poemId;//存在勘误的诗歌id
+	private long poemId;//存在勘误的诗歌id
 	private Timestamp verifyTime;//勘误审核时间
 	
 	public Timestamp getVerifyTime() {
@@ -56,7 +56,7 @@ public class ErrorInfo {
 	public void setUserId(int userId) {
 		this.userId = userId;
 	}
-	public int getPoemId() {
+	public long getPoemId() {
 		return poemId;
 	}
 	public void setPoemId(int poemId) {
